refactor(backend): move the API-key routes out of v1.py into their own feature module
v1.py is the 3,016-line residue of the pre-feature-module API layout. Its own
banner comments already mark eight sections; this moves the last of them —
the six /api/api-keys routes and the three request/response models they own —
into backend/api/features/api_keys/, alongside the other feature packages.

The published API is unchanged: the exported OpenAPI schema is byte-identical
before and after, and every one of the app's 357 operations still resolves to
the same handler.
